clean up, sorting, wpf-controlled custom showroom, etc

+++ AcTools.Render/Base/SceneRenderer.cs
@@ -13,7 +13,7 @@ namespace AcTools.Render.Base {
             Scene = new RenderableList();
         }
 
-        protected long GetTrianglesCount() {
+        protected int GetTrianglesCount() {
             return Scene.TrianglesCount;
         }
